doc(download): documentation
